pyproject.toml
[tool.poetry]
name = "phylo"
version = "0.1.0"
description = ""
authors = ["Onno Eberhard <onnoeberhard@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ncbi-datasets-pylib = "^12.27.1"
matplotlib = "^3.5.1"
jupyter = "^1.0.0"
numpy = "^1.22.1"
seaborn = "^0.11.2"
pandas = "^1.4.0"
python-Levenshtein = "^0.12.2"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
